animation added, market added

+++ Assets/Scripts/ScriptableObject/Customers.cs
@@ -12,6 +12,6 @@ public class ProductsCollect
 {
     public GameObject customers;
     public int productsCount;
-    public int money;
+    public float money;
 }
